% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_read_parse.R
\name{df_to_vector}
\alias{df_to_vector}
\title{Turn a data frame into vector}
\usage{
df_to_vector(.data)
}
\arguments{
\item{.data}{}
}
\description{
Use a data frame's first column to a vectors's names, the second column to its values.
}
