% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_study.R
\name{flatten_study}
\alias{flatten_study}
\title{Flatten a Study with a Recursive File Structure}
\usage{
flatten_study(origin, dest, delim = "_")
}
\arguments{
\item{delim}{}
}
\description{
Flatten a Study with a Recursive File Structure
}
