% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{flatten_study_dir}
\alias{flatten_study_dir}
\title{Flatten a Study with a Recursive File Structure}
\usage{
flatten_study_dir(origin, dest, delim = "_")
}
\arguments{
\item{origin}{origin path}

\item{dest}{destination path}

\item{delim}{file name delimiter}
}
\description{
Flatten a Study with a Recursive File Structure
}
