% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{lift_meta}
\alias{lift_meta}
\title{Lift Metadata from Workbooks in a Study}
\usage{
lift_meta(study)
}
\arguments{
\item{study}{a psyphr study}
}
\value{
a psyphr study
}
\description{
Lift Metadata from Workbooks in a Study
}
