% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{lift_study_dir}
\alias{lift_study_dir}
\title{Lift a Study with a Flat File Structure}
\usage{
lift_study_dir(origin, dest, delim = "_")
}
\arguments{
\item{origin}{a character string; origin path}

\item{dest}{a character string; destination path}

\item{delim}{a character string; file name delimiter}
}
\description{
Lift a Study with a Flat File Structure
}
