% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook.R
\name{print.psyphr_workbook}
\alias{print.psyphr_workbook}
\title{Print Brief Info on psyphr_workbook}
\usage{
\method{print}{psyphr_workbook}(x, ...)
}
\arguments{
\item{x}{a psyphr workbook}

\item{...}{dot-dot-dot}
}
\description{
Print Brief Info on psyphr_workbook
}
