% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_read_parse.R
\name{read_MW}
\alias{read_MW}
\title{Read a MindWare Workbook}
\usage{
read_MW(path)
}
\arguments{
\item{path}{file path to workbook}
}
\value{
a list of data frames, as a S3 object
}
\description{
Read a MindWare Workbook
}
