% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_read_parse.R
\name{read_MW_workbook}
\alias{read_MW_workbook}
\title{Read a MindWare Workbook in Excel format}
\usage{
read_MW_workbook(path, device_vendor = NULL)
}
\arguments{
\item{device_vendor}{}
}
\description{
Read a MindWare Workbook in Excel format
}
