% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{read_study}
\alias{read_study}
\title{Read a study from a directory}
\usage{
read_study(path, delim = "_")
}
\arguments{
\item{path}{path to a study directory}
}
\value{
a nested data frame including IDs
}
\description{
Read a study from a directory
}
