% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_hrv_editing_columns}
\alias{rename_hrv_editing_columns}
\title{Rename sheet columns after the sheet has been transformed.}
\usage{
rename_hrv_editing_columns(sheet)
}
\arguments{
\item{sheet}{}
}
\value{
a dataframe with renamed columns
}
\description{
Functions to change the column names to be easier to use by later analysis scripts.
They should only be called after transform_editing_sheet has already been used.
}
\details{
TODO: add error handling in case this is run before transform_editing_sheet has been run, and the
column names are not as expected. I.e., detect whether the second calumn name is a number, and if so,
exit out gracefully with a message to run transform_*_sheet function first.
}
\examples{
rename_hrv_editing_columns(sheet_data)

}
