% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{stash}
\alias{stash}
\title{Stash an R Object onto Disk}
\usage{
stash(object, dir_path = tempdir())
}
\arguments{
\item{dir_path}{}
}
\value{
a stash_pointer obj
}
\description{
Stash an R Object onto Disk
}
