% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_read_parse.R
\name{tidy_MW_EDA}
\alias{tidy_MW_EDA}
\title{Parsing and tidying a Mindware EDA workbook}
\usage{
tidy_MW_EDA(workbook)
}
\arguments{
\item{workbook}{}
}
\description{
Parsing and tidying a Mindware EDA workbook
}
