% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{transform_editing_sheet}
\alias{transform_editing_sheet}
\title{Process "Editing Stats" sheet}
\usage{
transform_editing_sheet(file_name, sheet)
}
\arguments{
\item{sheet}{}
}
\value{
a dataframe of segment data with a column containing the file name
}
\description{
The transform_editing_sheet function transforms the raw Editing Stats sheet to use the Segment Names as
column names, make all values numeric, and add the name of the file as a column, so the data includes
the source designation.
}
\examples{
transform_editing_sheet(name_of_file, sheet_data)

}
