% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_hrv_stats.R
\name{transform_hrv_sheet}
\alias{transform_hrv_sheet}
\title{Process "HRV Stats" sheet}
\usage{
transform_hrv_sheet(file_name, sheet, rename_columns = TRUE)
}
\arguments{
\item{rename_columns}{}
}
\description{
The transform_hrv_sheet function ...
}
\examples{

}
