% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{transform_settings_sheet}
\alias{transform_settings_sheet}
\title{Process "Settings" sheet}
\usage{
transform_settings_sheet(file_name, sheet)
}
\arguments{
\item{sheet}{}
}
\description{
The transform_settings_sheet function transforms the raw Settings sheet to "flip it on its side" so
variables are columns and values are rows
}
\details{
IMPORTANT NOTE: This needs to be tested
}
