% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree}
\alias{tree}
\title{Recursively Print out the Structure of a List}
\usage{
tree(x = .GlobalEnv, unclass = FALSE, compact = TRUE, rich = FALSE,
  all.names = FALSE)
}
\arguments{
\item{x}{a list, or another subsettable object}

\item{unclass}{expand list-like objects}

\item{compact}{don't show unnamed elements}

\item{rich}{show length & memory size of element}

\item{all.names}{show hidden objects in an environment}
}
\value{
character string of the screen output
}
\description{
Recursively Print out the Structure of a List
}
