% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{print.psyphr_study}
\alias{print.psyphr_study}
\title{Print a Summary of a Psyphr Study}
\usage{
\method{print}{psyphr_study}(x, ...)
}
\arguments{
\item{x}{a psyphr study object}

\item{...}{dot-dot-dot}
}
\description{
Print a Summary of a Psyphr Study
}
