% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook_read_parse.R
\name{read_MW}
\alias{read_MW}
\alias{read_MW_EDA}
\alias{read_MW_HRV}
\alias{read_MW_EMG}
\alias{read_MW_IMP}
\alias{read_MW_BPV}
\title{Read a MindWare Workbook}
\usage{
read_MW_EDA(path)

read_MW_HRV(path)

read_MW_EMG(path)

read_MW_IMP(path)

read_MW_BPV(path)
}
\arguments{
\item{path}{}
}
\value{
a list of Mindware data sheets
}
\description{
Read a MindWare Workbook
}
