% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{read_MW_study}
\alias{read_MW_study}
\title{Read a study from a directory}
\usage{
read_MW_study(path, structure = "flat", stash = FALSE,
  stash_dir_path = tempdir())
}
\arguments{
\item{path}{path to a study directory; a character string}

\item{structure}{directory structure; "flat" or "recursive"}

\item{stash}{whether to cache data on disk; logical}

\item{stash_dir_path}{cache directory; a character string}
}
\value{
a data frame; psyphr study S3 object
}
\description{
Read a study from a directory
}
