% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{read_study}
\alias{read_study}
\title{Read a study}
\usage{
read_study(path, ...)
}
\arguments{
\item{path}{path to a study directory; a character string}

\item{...}{dot-dot-dot}
}
\description{
Read a study
}
