% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{read_study}
\alias{read_study}
\title{Read a study from a directory}
\usage{
read_study(path)
}
\arguments{
\item{path}{a character string; path to a study directory}
}
\value{
a data frame; psyphr study S3 object
}
\description{
Read a study from a directory
}
