% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree}
\alias{tree}
\title{Recursively Print out the Structure of a List}
\usage{
tree(x = .GlobalEnv, limit = 5, class = TRUE, unclass = FALSE,
  compact = TRUE, rich = FALSE, all.names = FALSE)
}
\arguments{
\item{x}{a list or environment}

\item{limit}{recursive level limit}

\item{class}{show class of object}

\item{unclass}{expand list-like objects}

\item{compact}{don't show unnamed elements}

\item{rich}{show length & memory size of element}

\item{all.names}{show hidden objects in an environment}
}
\value{
character string of the screen output
}
\description{
Recursively Print out the Structure of a List
}
\references{
https://gist.github.com/iqis/926c189da89a32b9702c340c42401c67
}
